Sync with main

+++ Software/src/datalayer/datalayer_extended.h
@@ -425,7 +425,7 @@ struct DATALAYER_INFO_FORD_MACH_E {
   uint16_t pid_hvb_contactor_open_leak_resistance = 0;
   uint8_t pid_hvb_soh = 0;
   uint16_t pid_hvb_voltage = 0;
-  uint8_t pid_hvb_calendar_age_months = 0;
+  uint16_t pid_hvb_calendar_age_months = 0;
   uint16_t pid_battery_capacity_ah = 0;
   uint8_t pid_maintenance_rebalance_status = 0;
 };
